There are very minor performance improvements from this:

Before:
```
agb::agbabi::test::memcpy::test_all_of_memcpy...[ok: 53510906 c ≈ 3.19 s]
agb::agbabi::test::memcpy::test_all_of_memcpy4...[ok: 39371480 c ≈ 2.35 s]
agb::agbabi::test::memcpy::test_all_of_memcpy4_non_word_length...[ok: 45755219 c ≈ 2.73 s]
agb::agbabi::test::memcpy::test_memcpy4_with_different_sizes...[ok: 244971 c ≈ 0.01 s]
agb::agbabi::test::memset::test_memset4_with_different_sizes_and_offsets...[ok: 3326916 c ≈ 0.2 s]
agb::agbabi::test::memset::test_memset4_with_non_word_size_sizes_and_offsets...[ok: 10717390 c ≈ 0.64 s]
agb::agbabi::test::memset::test_memset_with_different_sizes_and_offsets...[ok: 2525106 c ≈ 0.15 s]
```

After:
```
agb::agbabi::test::memcpy::test_all_of_memcpy...[ok: 53494903 c ≈ 3.19 s]
agb::agbabi::test::memcpy::test_all_of_memcpy4...[ok: 39371450 c ≈ 2.35 s]
agb::agbabi::test::memcpy::test_all_of_memcpy4_non_word_length...[ok: 45755189 c ≈ 2.73 s]
agb::agbabi::test::memcpy::test_memcpy4_with_different_sizes...[ok: 243639 c ≈ 0.01 s]
agb::agbabi::test::memset::test_memset4_with_different_sizes_and_offsets...[ok: 3317581 c ≈ 0.2 s]
agb::agbabi::test::memset::test_memset4_with_non_word_size_sizes_and_offsets...[ok: 10699685 c ≈ 0.64 s]
agb::agbabi::test::memset::test_memset_with_different_sizes_and_offsets...[ok: 2512914 c ≈ 0.15 s]
```

- [x] No changelog update needed